refactor: Move utility methods
Move common utility methods into parent class: LinkedList.java in order to increase code readability and separate from operation logic in class: SingleLinkedList.java
